bot/requirements.txt
python-dotenv==1.2.1
python-telegram-bot==22.6
python-telegram-bot-calendar==1.0.5
requests==2.32.5